src/front-end, front-end/components: Integration of inline dropdowns

+++ src/front-end/localStorage/deleteFiles/deleteAuxiliarFunctions.js
@@ -32,32 +32,44 @@ export function handleContent (collection, callback) {
 }
 
 export function handleTrackers (collection, callback) {
-	let isNotRecurringTracker = !collection.recurringTrackers
-	let isRecurringTrackerEmpty = collection.recurringTrackers && collection.recurringTrackers.length === 0;
-	let doneRecurring = isNotRecurringTracker || isRecurringTrackerEmpty;
-	if (collection.trackers.length === 0 && doneRecurring) {
-		callback(null, collection);
-	} else {
-		localStorage.deleteTrackerByID(collection.trackers[0], false, (failedDeleteTracker) => {
-			/* istanbul ignore next */
-			if (failedDeleteTracker) {
-				/* istanbul ignore next */
-				callback(failedDeleteTracker, null);
-				/* istanbul ignore next */
-			} else if (collection.recurringTrackers && collection.recurringTrackers.length > 0) {
-					localStorage.deleteTrackerByID(collection.recurringTrackers[0], false, (failedDeleteRecurringTracker) => {
+	localStorage.readUser((failedRead, user) => {
+		if (failedRead) {
+			callback(failedRead);
+		} else {
+			collection = user[`${collection.objectType}s`].filter((reference) => reference.id === collection.id)[0];
+			let isNotRecurringTracker = !collection.recurringTrackers
+			let isRecurringTrackerEmpty = collection.recurringTrackers && collection.recurringTrackers.length === 0;
+			let doneRecurring = isNotRecurringTracker || isRecurringTrackerEmpty;
+			if (collection.trackers.length === 0 && doneRecurring) {
+				console.log("finished this");
+				callback(null, collection);
+			} else {
+				localStorage.deleteTrackerByID(collection.trackers[0], false, (failedDeleteTracker) => {
+					/* istanbul ignore next */
+					if (failedDeleteTracker) {
 						/* istanbul ignore next */
-						if (failedDeleteRecurringTracker) {
+						console.log("failed original");
+						callback(failedDeleteTracker, null);
+						/* istanbul ignore next */
+					} else if (collection.recurringTrackers && collection.recurringTrackers.length > 0) {
+						localStorage.deleteTrackerByID(collection.recurringTrackers[0], false, (failedDeleteRecurringTracker) => {
 							/* istanbul ignore next */
-							callback(failedDeleteRecurringTracker);
-							/* istanbul ignore next */
-						} else {
-							handleTrackers(collection, callback);
-						}
-					});
-				} else {
-					handleTrackers(collection, callback);
-				}
-		});
-	}
+							if (failedDeleteRecurringTracker) {
+								/* istanbul ignore next */
+								console.log("failed recurring");
+								callback(failedDeleteRecurringTracker, null);
+								/* istanbul ignore next */
+							} else {
+								console.log("handling next iteration");
+								handleTrackers(collection, callback);
+							}
+						});
+					} else {
+						console.log("handling next iteration without recurring")
+						handleTrackers(collection, callback);
+					}
+				});
+			}
+		}
+	})
 }
